Created a component called ActionButton to replace the StartButton, ResetButton, and StopButton components.
Created a test for the ActionButton component to check that it runs the function passed to it with props when it's clicked.

Still figuring out how to test the Clock class in clock.ts.

Refactored Clock class, it now runs a given function passed to it once every second.

Made two fields, the private onTick method and the private everyOneSecond field, readonly.

Refactored type imports to use the type keyword, explicitly stating their purpose as type-imports.

Deleted StartButton, ResetButton, and StopButton components and removed them from the StopWatch component; the ActionButton component has replaced it.

Set up the ActionButton components in the StopWatch component to start and stop a Clock instance.

Set up stop watch time state in the StopWatch component and passed the time to the ClockDisplay component to be formatted into hours:minutes:seconds.

Created a hook to memoize Clock instances, so that they will persist on component re-renders.