Added a new required prop to the ActionButton component called ariaControls, which specifies what element the button controls.
Added display of grid to the App component and positioned everything in the center.

Refactored AriaRoles.ts to ariaTypes.ts, so more types can be exported from it.

Created a new aria type for the aria-controls attribute.

Created an HTML_Types.ts file for HTML types.  Created an HTML type for HTML IDs.

Created a Sass mixins file for app default styles.  Created a button mixin to add app default button styles (removing grey background, border, etc.).

Created a Sass mixins file for reusable styles.  Created a mixin for rounding corners on elements.

Styled the ClockDisplay component, centering the text and increasing the font size.

Added an HTML ID to the ClockDisplay component, so it can be identified by other elements using the aria-controls attribute which control the ClockDisplay.

Removed tests for now deleted individual button components from the StopWatch.test.tsx file.  Replaced them with the new StopWatchButtons component, which renders all the ActionButton components for the StopWatch component.

Removed ActionButton components from the StopWatch component, replacing them with the new StopWatchButtons components, which renders all the ActionButtons for the StopWatch.

Created styles for the StopWatchButtons component, giving the containing div a grid layout and applying styles to the buttons inside it.

Created tests for the StopWatchButtons component to check if it renders each button (a start button, a stop button, and a reset button).