Applied more centering styles to the ClockDisplay component.
Created a new component, TimerBar, to visually display a measurement of the time associated with a certain clock function.

Created styles for the TimerBar component, making it a circle that can wrap around content.

Added a TimerBar component to the StopWatch component's markup.

Created a new hook for the TimerBar component to toggle and get its visibility.  Added this hook to the StopWatch component for its TimeBar component.

Added a grid layout to the StopWatch component, placing the ClockDisplay component inside the TimerBar component and placing the control buttons below them.

Removed the stopWatchComponentTitle constant.