Created a colors enum for hex code colors used in the TypeScript files.
Added some spacing between the ClockDisplay component and the StopWatchButtons component rendered in the StopWatch component.

Implemented functionality and styling for the TimerBar component.  It's a ring that fills in a circular fashion based on the time in seconds remaining out of the total passed to it.  It also has a nice gradient when the light theme is active and a simple white color when the dark theme is active.

Removed TimeBar component and its associated hook and functions from the StopWatch component since it will not make use of them; other components will.

Formatted the StopWatchButtons component's class name for readability.

Created two functions for the theme context to check if the theme is light or dark.

Refactored the time bar visibility feature to make use of intro and outro animations.

Added a temporary prop to the TimerBar component, so it can save the total time it decrements from as its maximum value (100% bar filled).