Added a transition to theme swapping.
Created a component, ThemeSwitch, which renders a button that switches the theme.  Gave it animations when clicked and when the knob changes positions.

Created a component, AppHeader, for housing components in a header.  Added a ThemeSwitch component to it and moved it to the end (the right) of the header.