Added more roles to the AriaRoles enum in ariaTypes.ts.  Switched to using these roles over string literal values in test files.
Removed clock.test.ts due to difficulty testing.

Created a new time type called TimeInMilliseconds.

Refactored the clock class' startClock function to accept options.  An option of precise will instruct the clock ticker to run every 1 millisecond instead of every 1 second.  Useful where that level of precision is necessary, such as in the StopWatch component.

Refactored the ClockDisplay component to use milliseconds, which allows it to work with greater precision when necessary.

Added a comment before the clockDisplayID export in ClockDisplay.tsx for clarification of its purpose.

Decreased the font size of the ClockDisplay component at smaller screen sizes to account for the possibility of displaying milliseconds.

Refactored the redundantly named stopStopWatch function in the StopWatch component to suspendStopWatch for improved readability.

Moved the resetStopWatchTime function closer to the other stop watch functions, startStopWatch and suspendStopWatch.

Refactored the props for the StopWatchButtons component to reduce confusion.

Fixed the tests in StopWatchButtons.test.tsx since they were broken due to relying on context that hadn't been provided when rendering them during the test cases.