Refactored the TimerBar component to be called TimerProgressBar.
Refactored the UseTimerBarVisibilityToggle to useVisibilityToggle.

Created a component for the background of the timer progress bar.

Created a new component for providing a gradient element to SVGs called
GradientProvider.

Simplified the TimerProgressBar component's markup by moving parts that
could be their own component into their own component files.  And
wrapping the progress bar in a GradientProvider component.