Created types for SVG stops passed as props to the GradientProvider component and for gradient color URLs.
Moved gradient type useState hook closer to where it's used.

Created another piece of state for the GradientProvider component to
store stops while they are processed for keying.

Created a useEffect hook to run a function that adds keys to the stops
provided to the GradientProvider component if they are provided as an
array.