Created the EditableClockDisplayOverlay component and its styles.
Created state for enabling/disabling editing.

When editing, an input will be rendered in place of the normal
ClockDisplay component.

Added some functions to format the time in hours:minutes:seconds.

Set up component to disable editing and render the normal ClockDisplay
component when it updates its parent's time
with the function passed to it, updateTimeInSeconds.