Added functionality and state to the Timer component.
Added two parts of state, one for the time to count from, which the user
can update and one for decrementing the time, that the timer uses.

Added a useEffect hook to set the counter time when the user updates the
time to count from.

Added a useEffect hook to prevent negative time when counting down.

Added state to keep track of if the clock is started.