fixup! Created a new test for the StopWatch component to test only rendering the Laps component when the stop watch is started.
Changed the conditional controlling whether the Laps component was rendered to use the value that determines if the stop watch is active rather than if it's started.  Otherwise, the laps will not show if the stop watch is stopped but not reset.

Refactored the afterClockStarts boolean to whileStopWatchActive which a much clearer and less redundant name.