Fixed issue with the ToggleSwitch component's knob.
Applied a grid layout to the toggle switch's button to keep the knob inside it when the button scales.  Also refactored the margin, so that negative values were not used.

Additionally improved the accuracy of the margin to pixel-perfection by utilizing the knob's 1:2 size ratio compared to the button, applying a left margin of half the size of the button to keep the knob in its container.

Refactored the toggle switch's scaling to use the CSS transform scale which provides more consistency when the component is resized.