Refactored ToggleSwitch scaling.
Moved the scaling logic from the SCSS file into the component file in order to allow the component to be scaled by passing a prop to it.

Refactored the scaling so that it affects the layout instead of ignoring it.

Created a function which takes a number and makes it an em unit.