Added icons to the clock feature links.
Replaced the placeholder text for each clock feature link with an icon representing them.

Added a grid layout and spaced each link apart.

Added a dark theme class to invert the icon colors when the dark theme is applied, so that they are visible and comply with the theme.

Passed the boolean representing if the theme was dark from the AppHeader component down to the ClockFeatureLinks component, so that it can use it without having to re-grab the theme context.

Replaced the timer icon, so it stands out from the stop watch icon.