Added more styling to the clock feature links.
Added styling to hide the links to other pages by default and show them when the component is hovered over or focused.

Placed the current page link at the beginning of the grid layout.

Created logic to assign a class name to the li wrapper for the link component that corresponds to the current URL path.