Implemented transition and refactored styling.
Refactored the route map into an object as Map was not a necessary data structure.

Created a transition for sliding the non-active links in and out of the active one.

Moved routes into state in order to update the link order in the markup, allowing the current page link to always be the first in order.

Refactored the component styles, reducing the complexity.

Added new, resized stop watch and timer icons.