Created a component for the clock feature link icons.
Renamed icons folder to accommodate the component and created a new folder for the icons in the ClockFeatureLinks components directory.

Moved JSX related to the icons for the clock features into its own component.

Moved styles related to the icons to a stylesheet for the new component.

Passed necessary props to the new component.