Implemented 12h/24h toggling; TimeClock.tsx.
Used the use12hTime hook in the TimeClock component to allow it to toggle between 24h and AM/PM time format by clicking the ToggleSwitch component that it renders.

Created a constant for the amount of afternoon hours in milliseconds, so time passed to the ClockDisplay component would be adjusted depending on if AM/PM was toggled on.