Set up sliding logic for the SlidingRouteMenu component.
Created a new prop for the SlidingRouteMenuItem component to pass a custom style object in order to apply the slide translation to each item from the SlidingRouteMenu component.

Added sliding to the SlidingRouteMenuItem components by passing a custom style attribute to each item as getting the amount to slide it by calling a function which returns how many ems to translate the menu item.

Adjusted the grid layout for the SlidingRouteMenu, so that the gaps are consistent with the default icon size (1.25em or 20px).

Applied a transition to the SlidingRouteMenuItems, so that their sliding behaviour is animated.

Added a white background to each SlidingRouteMenuItem and created a class for the dark theme variant.