Added a new feature to the ToggleSwitch component.
Created a test for rendering a component from the ToggleSwitch that is passed the toggle state, so that the information regarding what the toggle state means is displayed.

Created an optional prop for the ToggleSwitch component for passing the toggle state to a component to be rendered by the ToggleSwitch.

Passed the toggle state to the optional prop in the markup, if it's provided.