Refactored the ClockDisplay component.
Added local state to the component, so that it is in charge of its controlled input and only submits the time to the outer component once the input field's focus has been blurred.

This allows the time entered to remain accurate instead of auto-converting it while it's being typed, resulting in unintended behaviour.

Removed the unused FormattedTimeUnit type from timeTypes.ts.