Added a ThemeIcon component to AppHeader.tsx.
Added the ThemeIcon component to the markup and wrapped it and the ToggleSwitch component in a div to apply layout styles.

Created the layout styles in the AppHeader style module.

Added a media query to the ThemeIcon style module to increase the icon size at larger screens.

Fixed wrong class name used in the media query in the ToggleSwitch style module.