Added an element displaying the time format, TimeClock.tsx.
Created two tests for the TimeClock component for displaying text regarding the current time format, 24H or AM/PM (12H).

Added an element next to the ToggleSwitch component in the TimeClock component's markup and wrapped them in a div and added a layout style.  The element displays which time format is being used, either 24H or AM/PM (12H).

Created a grid layout for the time format area which contains the toggle switch and the time format text.  Centered the content and limited the width according to the size of the ClockDisplay component that is rendered above it.

Set up styles for the layout and the time format text at larger screen sizes, increasing the text font size and readjusting width of the layout.

Fixed a test in TimeClock.test.tsx which tests the toggle switch changing the time format to 12H.