Vastly simplified code setting the display time; ClockDisplay component.
Tested for valid input when editing the ClockDisplay component.

Simplified the logic in validating and updating the display time for the ClockDisplay component, removing two functions from clockDisplayUtils.ts and reducing the complexity of the setDisplayTime function.

Created a new string utility function that checks if a string is a number.  Used this utility in validating the time entered in the ClockDisplay component.